wait for receipt
